+++ CommonMark/Parser/Blocks/Delimiters/RomanListItemHandler.cs
@@ -70,7 +70,7 @@ namespace CommonMark.Parser.Blocks.Delimiters
         /// <param name="value">Current character value.</param>
         /// <param name="curChar">Current character.</param>
         /// <returns><c>true</c> if successful.</returns>
-        protected override bool AdjustStart(ref int start, ref int value, char curChar)
+        protected override bool AdjustStart(ref int start, ref short value, char curChar)
         {
             var prevValue = value;
 
@@ -83,7 +83,7 @@ namespace CommonMark.Parser.Blocks.Delimiters
                 {
                     if (prevValue * 2 > value)
                         return false;
-                    value -= prevValue * 2;
+                    value -= (short)(prevValue * 2);
                 }
                 else if (value > start)
                     return false;
